Move AesEncryptLayer to samples
